+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - DF-811 Add support for upsert
 - DF-895 Added support for username based authentication
+- DF-1084 added support for Admin User email invites
 ### Changed
 - DF-1028 Ignored JWT explicitly when making login requests
 - Moved database configuration to service and the new NoDbModel-based handler